Main dev

+++ back/db/operaciones/empleados/modificar_db.py
@@ -41,17 +41,20 @@ def modificar_empleado(
     dni_final = empleado_dni if (dni_nuevo is None) else dni_nuevo
     nombre_final = nombre if nombre is not None else datos_actuales["nombre"]
     apellido_final = apellido if apellido is not None else datos_actuales["apellido"]
-    correo_final = correo if correo is not None else datos_actuales["correo"]
     genero_final = genero if genero is not None else datos_actuales["genero"]
     rol_id_final = rol_id if rol_id is not None else datos_actuales["rol_id"]
     
+    # CONTROL AD-HOC: Si no es un profesor y su correo es el mismo, se va a recibir en
+    # el modificar_empleado_service.
+
     query_update = f"""
         UPDATE Usuario
         SET dni = {dni_final},
             nombre = '{nombre_final}',
-            apellido = '{apellido_final}',
-            correo = '{correo_final}',
-            genero = '{genero_final}',
+            apellido = '{apellido_final}',"""
+    if correo is not None:
+        query_update += f"correo = '{correo}',"
+    query_update += f"""genero = '{genero_final}',
             rol_id = {rol_id_final}
         WHERE dni = {empleado_dni}
     """
@@ -117,18 +120,20 @@ def borrar_empleado(empleado_dni: int, cursor) -> dict:
     usuario = ejecutar_fetchone(query_verificacion, cursor)
     print("IMPRIMIENDO RESPUESTA DEL FETCHONEEEEEEEEEEEEEEEEEEEEEEEEEEEEE",usuario)
 
-    if not usuario:
+    if usuario.get("status") == "error" or not usuario.get("data"):
         return {
             "status": "error",
             "message": "Empleado no encontrado"
         }
-    elif usuario['data']["rol_id"] not in (0, 1, 2, 5): # o sea si el usuario no es ni gerente ni administrador ni profesor
+    rol_actual = usuario["data"]["rol_id"]
+
+    # FIX 2,  Agarrar si ya estaba eliminado
+    if rol_actual == 4:  # o sea si el usuario ya estaba eliminado
         return {
             "status": "error",
-            "message": "El usuario no es un empleado"
+            "message": "El empleado ya estaba eliminado"
         }
-    elif usuario["data"]['rol_id'] == 5: # o sea si el usuario es un profesor
-        ## Aca hay que mejorar la logica y verificar si el profesor tiene clases a su cargo, y en ese caso no permitir el borrado
+    elif rol_actual == 5: # o sea si el usuario es un profesor
         print("El usuario es un profesor, se procede a verificar si tiene clases a su cargo")
         profesor_id = usuario["data"]["id"]
         query_clases_profesor = f"""
@@ -146,16 +151,35 @@ def borrar_empleado(empleado_dni: int, cursor) -> dict:
                 "message": f"No se puede eliminar al profesor. Tiene {resultado_clases['data']['total_clases']} clase/s asignada/s"
             }
 
-    
     # Borrado lógico: se modifica el rol a 4 (eliminado) conservando los datos personales
     # AHORA ES BORRADO FISICO
     # SI QUEREMOS HACER BORRADO TAMBIÉN EN SUS REFERENCIAS (FOREIGN KEYS), HAY QUE UTILIZAR DELETE ON CASCADE. PERO POR LAS DUDAS LO DEJO COMO DELETE NORMAL
-    query_delete = f"""
-        DELETE FROM Usuario
-        WHERE dni = {empleado_dni}
-    """
+    # query_delete = f"""
+    #     DELETE FROM Usuario
+    #     WHERE dni = {empleado_dni}
+    # """
+    # AHORA ES LÓGICO LA RREPUTAMADRE DECIDANSÉ MATERIA DE LA REVERENDA CONCHA
+    # Borrado lógico: se modifica el rol a 4 (eliminado) conservando los datos personales
+    # AHORA PARECE QUE EN REALIDAD EL FÍSICO ES PARA LOS CLIENTES. ME QUIEREN VOLVER LOCO
 
-    return ejecutar_query(query_delete, cursor)
+    # BIFURCACIÓN DE BORRADO (FÍSICO VS LÓGICO)
+    # Si NO es un empleado (0, 1, 2 = admins/gerentes, 5 = profe), asumimos que es CLIENTE
+    if rol_actual not in (0, 1, 2, 5): 
+        # BORRADO FÍSICO
+        query_delete = f"""
+            DELETE FROM Usuario
+            WHERE dni = {empleado_dni}
+        """
+        return ejecutar_query(query_delete, cursor)
+    else:
+        # BORRADO LÓGICO (Es un empleado)
+        query_update = f"""
+            UPDATE Usuario
+            SET rol_id = 4
+            WHERE dni = {empleado_dni}
+        """
+        return ejecutar_query(query_update, cursor)
+
 
 def desactivar_empleado(empleado_dni: int, cursor) -> dict:
     """Desactiva un empleado específico de la base de datos, utilizando su DNI como referencia.
